requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.1
flask-restplus==0.13.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
Flask-Testing==0.8.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
mistune==3.0.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.3
azure-functions

